+++ examples/bumpy_surface_noise.rb
@@ -10,9 +10,7 @@ attr_reader :srface, :box2d, :particles
 def setup
   sketch_title 'Bumpy Surface Noise'  
   # Initialize box2d physics and create the world
-  @box2d = Box2D.new(self)
-  box2d.init_options(gravity: [0, -20])
-  box2d.create_world
+  @box2d = WorldBuilder.build(app: self, gravity: [0, -20])
   # to later set a custom gravity
   # box2d.gravity([0, -20])
   # Create the empty list