+++ examples/liquid_fun_test.rb
@@ -17,9 +17,7 @@ end
 
 def setup
   sketch_title 'Liquid Fun Test'
-  @box2d = Box2D.new(self)
-  box2d.init_options(gravity: [0, -10])
-  box2d.create_world
+  @box2d = WorldBuilder.build(app: self)
   @boxes = []
   box2d.world.set_particle_radius(0.15)
   box2d.world.set_particle_damping(0.2)