+++ examples/revolute_joint/revolute_joint.rb
@@ -16,8 +16,7 @@ end
 
 def setup
   sketch_title 'Revolute Joint'
-  @box2d = Box2D.new(self)
-  box2d.createWorld
+  @box2d = WorldBuilder.build(app: self)
   @windmill = Windmill.new(width / 2, 175)
   @system = ParticleSystem.new
 end